backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
faster-whisper
demucs
moviepy
ffmpeg-python
soundfile
pydantic
pydantic-settings
python-dotenv
librosa
numpy
torch
torchaudio
